Feature/ specification-compliant Float (#559)
* gql

* updATE CHANGELOG

* UPDATE CONCATENATE

* update

* update

* Double

* wrapped

+++ test/Feature/Input/Scalar/API.hs
@@ -31,7 +31,7 @@ testRes Arg {value} = pure value
 
 -- resolver
 data Query m = Query
-  { testFloat :: Arg Float -> m Float,
+  { testFloat :: Arg Double -> m Double,
     testInt :: Arg Int -> m Int,
     testString :: Arg Text -> m Text
   }
